Macros are too well structured ... please read.
By making the presence of the body the definition of a macro,
we are being quite limiting. The definition of a macro is
the presence of undelimimted expressions.

Whether undelimited expressions create a valid macro is another
thing, i don't see a reason why we can't just lop undelimited
expressions into a macro form which is just a list...

i mean it's the only the same as a body... which was represented
as a component of a macro form prior.