We're turning infix to targetted sends too prematurely.
Now that we have braced forms, we can just delay it and
let macroexpansion deal with it.

So to make that very explicit, since now `foo + bar`
is a macroForm, during macroexpansion when we find
a macro that doesn't make sense we can just let
the expander find infix stuff and shit out something
proper.